// src/gno/types.gen.ts
const crypto_PubKey = t.interface("crypto.PubKey");
// ...
const std_VestingSchedule = t.struct("std.VestingSchedule", {
  originalVesting: t.field(reprs.Coins, { json: "original_vesting", omitEmpty: true }),
  startTime: t.field(t.int64, { json: "start_time", omitEmpty: true }),
  endTime: t.field(t.int64, { json: "end_time", omitEmpty: true }),
  type: t.field(t.string, { omitEmpty: true }),
});

const std_BaseAccount = t.struct("std.BaseAccount", {
  address: reprs.Address,
  coins: reprs.Coins,
  pubKey: t.field(crypto_PubKey, { json: "public_key" }),
  accountNumber: t.field(t.uint64, { json: "account_number" }),
  sequence: t.uint64,
  vesting: t.field(std_VestingSchedule, { omitEmpty: true }),
});

const std_BaseSessionAccount = t.struct("std.BaseSessionAccount", {
  baseAccount: t.field(std_BaseAccount, { json: "BaseAccount" }),
  masterAddress: t.field(reprs.Address, { json: "master_address" }),
  expiresAt: t.field(t.int64, { json: "expires_at" }),
  spendLimit: t.field(reprs.Coins, { json: "spend_limit", omitEmpty: true }),
  spendPeriod: t.field(t.int64, { json: "spend_period", omitEmpty: true }),
  spendUsed: t.field(reprs.Coins, { json: "spend_used", omitEmpty: true }),
  spendReset: t.field(t.int64, { json: "spend_reset", omitEmpty: true }),
});

const std_Coin = reprs.Coin;

const std_GasPrice = t.struct("std.GasPrice", {
  gas: t.int64,
  price: std_Coin,
});

const std_Msg = t.interface("std.Msg");

const std_Fee = t.struct("std.Fee", {
  gasWanted: t.field(t.int64, { json: "gas_wanted" }),
  gasFee: t.field(std_Coin, { json: "gas_fee" }),
});

const std_Signature = t.struct("std.Signature", {
  pubKey: t.field(crypto_PubKey, { json: "pub_key" }),
  signature: t.bytes,
  sessionAddr: t.field(reprs.Address, { json: "session_addr", omitEmpty: true }),
});

const std_Tx = t.struct("std.Tx", {
  msgs: t.field(t.slice(std_Msg), { json: "msg" }),
  fee: std_Fee,
  signatures: t.slice(std_Signature),
  memo: t.string,
});

const std_MemFile = t.struct("std.MemFile", {
  name: t.string,
  body: t.string,
});

const std_MemPackage = t.struct("std.MemPackage", {
  name: t.string,
  path: t.string,
  files: t.slice(t.pointer(std_MemFile)),
  type: t.field(t.interface(), { omitEmpty: true }),
  info: t.field(t.interface(), { omitEmpty: true }),
});

const std_InternalError = t.struct("std.InternalError", {});

const std_TxDecodeError = t.struct("std.TxDecodeError", {});

const std_InvalidSequenceError = t.struct("std.InvalidSequenceError", {});

const std_UnauthorizedError = t.struct("std.UnauthorizedError", {});

const std_InsufficientFundsError = t.struct("std.InsufficientFundsError", {});

const std_UnknownRequestError = t.struct("std.UnknownRequestError", {});

const std_InvalidAddressError = t.struct("std.InvalidAddressError", {});

const std_UnknownAddressError = t.struct("std.UnknownAddressError", {});

const std_InvalidPubKeyError = t.struct("std.InvalidPubKeyError", {});

const std_InsufficientCoinsError = t.struct("std.InsufficientCoinsError", {});

const std_InvalidCoinsError = t.struct("std.InvalidCoinsError", {});

const std_InvalidGasWantedError = t.struct("std.InvalidGasWantedError", {});

const std_OutOfGasError = t.struct("std.OutOfGasError", {});

const std_MemoTooLargeError = t.struct("std.MemoTooLargeError", {});

const std_InsufficientFeeError = t.struct("std.InsufficientFeeError", {});

const std_TooManySignaturesError = t.struct("std.TooManySignaturesError", {});

const std_NoSignaturesError = t.struct("std.NoSignaturesError", {});

const std_GasOverflowError = t.struct("std.GasOverflowError", {});

const std_RestrictedTransferError = t.struct("std.RestrictedTransferError", {});

const std_SessionExpiredError = t.struct("std.SessionExpiredError", {});

const std_SessionNotFoundError = t.struct("std.SessionNotFoundError", {});

const std_SessionLimitError = t.struct("std.SessionLimitError", {});

const std_SessionNotAllowedError = t.struct("std.SessionNotAllowedError", {});

const std_InvalidVestingScheduleError = t.struct("std.InvalidVestingScheduleError", {});

const std_VestingLockedCoinsError = t.struct("std.VestingLockedCoinsError", {});

const sdk_Result = t.struct("sdk.Result", {
  responseBase: t.field(abci_ResponseBase, { json: "ResponseBase" }),
  gasWanted: t.field(t.int64, { json: "GasWanted" }),
  gasUsed: t.field(t.int64, { json: "GasUsed" }),
});

const params_Param = reprs.Param;
// ...
const vm_MsgCall = t.struct("vm.MsgCall", {
  caller: reprs.Address,
  send: reprs.Coins,
  maxDeposit: t.field(reprs.Coins, { json: "max_deposit" }),
  pkgPath: t.field(t.string, { json: "pkg_path" }),
  func: t.string,
  args: t.field(t.slice(t.string), { omitEmpty: true }),
});

const vm_MsgRun = t.struct("vm.MsgRun", {
  caller: reprs.Address,
  send: reprs.Coins,
  maxDeposit: t.field(reprs.Coins, { json: "max_deposit" }),
  package: t.pointer(std_MemPackage),
});

const vm_MsgAddPackage = t.struct("vm.MsgAddPackage", {
  creator: reprs.Address,
  package: t.pointer(std_MemPackage),
  send: reprs.Coins,
  maxDeposit: t.field(reprs.Coins, { json: "max_deposit" }),
});

const vm_MsgEnablePackage = t.struct("vm.MsgEnablePackage", {
  approver: reprs.Address,
  pkgPath: t.field(t.string, { json: "pkg_path" }),
  pkgHash: t.field(t.string, { json: "pkg_hash" }),
  pkgHeight: t.field(t.int64, { json: "pkg_height" }),
});

const vm_MsgRejectPackage = t.struct("vm.MsgRejectPackage", {
  sender: reprs.Address,
  pkgPath: t.field(t.string, { json: "pkg_path" }),
});

const vm_InvalidPkgPathError = t.struct("vm.InvalidPkgPathError", {});

const vm_NoRenderDeclError = t.struct("vm.NoRenderDeclError", {});

const vm_PkgExistError = t.struct("vm.PkgExistError", {});

const vm_InvalidStmtError = t.struct("vm.InvalidStmtError", {});

const vm_InvalidExprError = t.struct("vm.InvalidExprError", {});

const vm_TypeCheckError = t.struct("vm.TypeCheckError", {});

const vm_UnauthorizedUserError = t.struct("vm.UnauthorizedUserError", {});

const vm_InvalidPackageError = t.struct("vm.InvalidPackageError", {});

const vm_InvalidFileError = t.struct("vm.InvalidFileError", {});

const vm_ObjectNotFoundError = t.struct("vm.ObjectNotFoundError", {});

const vm_ExportSizeExceededError = t.struct("vm.ExportSizeExceededError", {});

const vm_ExportDepthExceededError = t.struct("vm.ExportDepthExceededError", {});

const vm_UnobservedSendError = t.struct("vm.UnobservedSendError", {});

const vm_UnspendableSendError = t.struct("vm.UnspendableSendError", {});

const vm_Params = t.struct("vm.Params", {
  sysNamesPkgPath: t.field(t.string, { json: "sysnames_pkgpath" }),
  sysCLAPkgPath: t.field(t.string, { json: "syscla_pkgpath" }),
  chainDomain: t.field(t.string, { json: "chain_domain" }),
  defaultDeposit: t.field(t.string, { json: "default_deposit" }),
  storagePrice: t.field(t.string, { json: "storage_price" }),
  storageFeeCollector: t.field(reprs.Address, { json: "storage_fee_collector" }),
  minGetReadDepth100: t.field(t.int64, { json: "min_get_read_depth_100" }),
  minSetReadDepth100: t.field(t.int64, { json: "min_set_read_depth_100" }),
  minWriteDepth100: t.field(t.int64, { json: "min_write_depth_100" }),
  fixedGetReadDepth100: t.field(t.int64, { json: "fixed_get_read_depth_100" }),
  fixedSetReadDepth100: t.field(t.int64, { json: "fixed_set_read_depth_100" }),
  fixedWriteDepth100: t.field(t.int64, { json: "fixed_write_depth_100" }),
  iterNextCostFlat: t.field(t.int64, { json: "iter_next_cost_flat" }),
  preprocessGasPerByte: t.field(t.int64, { json: "preprocess_gas_per_byte" }),
  codeSubmissionPolicy: t.field(t.string, { json: "code_submission_policy" }),
  codeSubmitters: t.field(t.slice(reprs.Address), { json: "code_submitters" }),
  pkgApprovers: t.field(t.slice(reprs.Address), { json: "pkg_approvers" }),
  runSubmitters: t.field(t.slice(reprs.Address), { json: "run_submitters" }),
  inertSubmissionCharge: t.field(t.string, { json: "inert_submission_charge" }),
  inertChargeCollector: t.field(reprs.Address, { json: "inert_charge_collector" }),
});

const vm_GenesisState = t.struct("vm.GenesisState", {
  params: vm_Params,
  realmParams: t.field(t.slice(params_Param), { json: "realm_params" }),
});
// ...
export const std = {
  BaseAccount: std_BaseAccount,
  BaseSessionAccount: std_BaseSessionAccount,
  Coin: std_Coin,
  Fee: std_Fee,
  GasOverflowError: std_GasOverflowError,
  GasPrice: std_GasPrice,
  InsufficientCoinsError: std_InsufficientCoinsError,
  InsufficientFeeError: std_InsufficientFeeError,
  InsufficientFundsError: std_InsufficientFundsError,
  InternalError: std_InternalError,
  InvalidAddressError: std_InvalidAddressError,
  InvalidCoinsError: std_InvalidCoinsError,
  InvalidGasWantedError: std_InvalidGasWantedError,
  InvalidPubKeyError: std_InvalidPubKeyError,
  InvalidSequenceError: std_InvalidSequenceError,
  InvalidVestingScheduleError: std_InvalidVestingScheduleError,
  MemFile: std_MemFile,
  MemPackage: std_MemPackage,
  MemoTooLargeError: std_MemoTooLargeError,
  Msg: std_Msg,
  NoSignaturesError: std_NoSignaturesError,
  OutOfGasError: std_OutOfGasError,
  RestrictedTransferError: std_RestrictedTransferError,
  SessionExpiredError: std_SessionExpiredError,
  SessionLimitError: std_SessionLimitError,
  SessionNotAllowedError: std_SessionNotAllowedError,
  SessionNotFoundError: std_SessionNotFoundError,
  Signature: std_Signature,
  TooManySignaturesError: std_TooManySignaturesError,
  Tx: std_Tx,
  TxDecodeError: std_TxDecodeError,
  UnauthorizedError: std_UnauthorizedError,
  UnknownAddressError: std_UnknownAddressError,
  UnknownRequestError: std_UnknownRequestError,
  VestingLockedCoinsError: std_VestingLockedCoinsError,
  VestingSchedule: std_VestingSchedule,
};

export const sdk = {
  Result: sdk_Result,
};

export const params = {
  Param: params_Param,
};
// ...
export const vm = {
  ExportDepthExceededError: vm_ExportDepthExceededError,
  ExportSizeExceededError: vm_ExportSizeExceededError,
  GenesisState: vm_GenesisState,
  InvalidExprError: vm_InvalidExprError,
  InvalidFileError: vm_InvalidFileError,
  InvalidPackageError: vm_InvalidPackageError,
  InvalidPkgPathError: vm_InvalidPkgPathError,
  InvalidStmtError: vm_InvalidStmtError,
  MsgAddPackage: vm_MsgAddPackage,
  MsgCall: vm_MsgCall,
  MsgEnablePackage: vm_MsgEnablePackage,
  MsgRejectPackage: vm_MsgRejectPackage,
  MsgRun: vm_MsgRun,
  NoRenderDeclError: vm_NoRenderDeclError,
  ObjectNotFoundError: vm_ObjectNotFoundError,
  Params: vm_Params,
  PkgExistError: vm_PkgExistError,
  TypeCheckError: vm_TypeCheckError,
  UnauthorizedUserError: vm_UnauthorizedUserError,
  UnobservedSendError: vm_UnobservedSendError,
  UnspendableSendError: vm_UnspendableSendError,
};